move component to right folder
